book-store-search-engine/requirements.txt
dotenv
requests
psycopg[binary]
transformers
sentence-transformers
qdrant-client[fastembed] 
fastembed
langchain
langchain-openai
grpcio
grpcio-tools
grpcio-health-checking
redis
einops
pandas
# Langsmith for pipeline tracing
langsmith
